Change PlatformDetection.GetFrameworkVersion() to use registry key value instead of minor build number
