+++ src/CoreFx.Private.TestUtilities/src/System/PlatformDetection.NonNetFx.cs
@@ -9,5 +9,8 @@ namespace System
     public static partial class PlatformDetection
     {
         public static bool TargetsNetFx452OrLower => false;
+        public static bool IsNetfx462OrNewer() { return false; }
+        public static bool IsNetfx470OrNewer() { return false; }
+        public static bool IsNetfx471OrNewer() { return false; }
     }
 }
